JobRequestFireStore.java
This is a work in progress. The file is not complete and testing needs to be done. The file requires both ClientHomeActivity and the DasherHomeActivity and the functions needs to be mapped to a button or function calls in the application. If there is something fundamentally wrong with the code let me know asap. The code has comments that tell what each function does along with the expected outcome when the function is called.