Add files via upload
Everything is Changed. The getinProgressJob function is having problems returning a JobRequest object. Can someone help me with that. Other than that the usual still applies, the functions are going to toast wether or not the function has succeeded or failed and the toast is set to clienthomeactivity or the dasherhomeactivity this may need to be changed.