Add files via upload
Able to return the JobRequest Instance. Also changed writeJobRequest to write to the dorm collection instead of inProgressJobs function. Made a function that moves the jobs from dorms collection into the inProgressJobs collection. onAssignmentAccept function was created it takes a job and changes the assigned timestamp, and also calls the movesToInProgressJobs function.